Move components to nested 'ClusterList' folder
